Add Kick login API and shared helper
