+++ scripts/login_kick.py
@@ -1,22 +1,5 @@
 import argparse
-import requests
-
-LOGIN_URL = "https://kick.com/api/v1/login"
-
-
-def login(email: str, password: str) -> str:
-    """Attempt to log in and return auth_token."""
-    session = requests.Session()
-    resp = session.post(
-        LOGIN_URL,
-        json={"email": email, "password": password},
-        headers={"User-Agent": "Mozilla/5.0"},
-    )
-    resp.raise_for_status()
-    token = session.cookies.get("auth_token") or resp.json().get("token")
-    if not token:
-        raise RuntimeError("auth_token not found in response")
-    return token
+from shared.kick import login
 
 
 def main() -> None:
